requirements.txt
# requirements.txt

# Ajoutez ici toutes les dépendances nécessaires, par exemple :
python-gitlab           # API GitLab
requests                # Requêtes HTTP (SonarQube, Dependency Track, DefectDojo)
python-dotenv           # Gestion des fichiers .env
pandas                  # Manipulation de données
openpyxl                # Export Excel
PyYAML                  # Lecture/écriture YAML
defectdojo-api          # Client DefectDojo (non officiel)
# dependency-track-api-client # Client Dependency Track (si utilisé, sinon requests)
tqdm                    # Barre de progression pour ETL
tabulate                # Affichage tabulaire (optionnel)
dateparser              # Parsing flexible des dates
python-dateutil         # Manipulation avancée des dates
pytest                  # Tests unitaires
pytest-cov              # Couverture de tests
apache-airflow[postgres]==2.7.3           # Orchestration ETL avec Airflow + support Postgres
Flask-Session==0.4.0           # Version compatible avec Airflow 2.7.x
Flask==2.2.5                   # Version compatible avec Airflow 2.7.x
# Veillez à NE PAS inclure python-daemon sur Windows